requirements.txt
# Core
numpy>=1.26.0
pillow>=10.0.0
tqdm>=4.66.0

# Vision / Embeddings (Torch + Torchvision: pick CPU or GPU wheel as per your system)
torch>=2.3.0
torchvision>=0.18.0

# Optional (for plotting/visualization from CLI)
matplotlib>=3.8.0

# MongoDB support
pymongo>=4.6.0

# Web application
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Environment variables
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
aiofiles>=23.2.1
